Display test name (#5)
* pageload test

create cypress test suite for test that are ran on the inital page load.

* html test

create cypress suite that executes a quick run of the html test of the
application.

* config react route

install the `react-router` library and config a base setup for hte
application. Root page can render list of test, and configured directory
to read exports from the `routes` dir.

* Test Route

Create route for the /:test route, use a conditional to render the
`Outlet` only when the user navigates to a test. This keeps the theme
toggle on the page at all times.

* test question route

Create route component that initializes test data to be used for
retrieving question and test information in the routes within. Currenly
a basic `Question` component can be rendered in the defined routes.

* Render test question

Use the `useParam` hook in `TestQuestion` to get the currnet question
number, get the state from the context and extract the question data
that is passed into the `Question` component to render the question.

* question switch

Apply the `useNavigate` hook to navigate to the next question once it
has been answered, pass in the URl to navigate to into the `Question`
component and reset the state.

* update context redcuer

Create and pass in a function that updates the context reducer when a
answer is submitted, it only updates the answer sheet which will be used
to display the `result` page.

* navigate function

Instead of calling `useNavigate` in the `Question` component, in
`TestQuestion` we create a functino that generate the `navigate` to
call, either going to the next question, or routing to the result page
with the answer sheet passed in.

* render results

`/result` can be navigated to and display a user's results after
completing a quiz.

* question off by one

adjust question number to ensure the first question from the dataset is
not skipped.

* navigate back to home

when test is completed, user can navigate to the root page to take
another quiz.

* save context

on a question submission, save the answer sheet into the local storage.

* state save on refresh

save the answersheet into local storage, so that if a user refreshes the
page, the answers saved onto that point are re-set into the context
through a `useEffect`.

* testing env

configure vitest for the project, and set `jsdom` to be used by default
for testing components.

* setup test config and create test

install the react testing library along with `user-event` to create a
util functino that renders a componentn and provides the `user-event`
methods as well. Modify vite config to provide absolute file paths and
modify eslint to ignore the `test` directory.

* `Question Test`

Configure and create multiple test for `Question` component, test ensure
that the component renders, displays the question, and can indicate
correct and incorrect options.

* 0.1.0

* render test name

set a state that displays the test name for the the currently selected
test.

* persist test name after refresh

set the test name in the local storage, so during mid-test, if the page
is refreshed, the test name can persist along with other data.