Validate url navigation (#7)
* current question count

Add a property to the reducer that keeps track of the question the user
should be on. Add the count to the local storage and ensure to remove it
on page load and when entering the results page.

* next question

modify `currentQuestion` to instead become the `nextQuestion`. It was
redundant to have a current question in the reducer since the question
number from the url served as an indicator of the "current" question.

* reducer answer sheet

Set the type of the answer sheet to either be null or boolean, on the
array initialization all indexes are set to null and changed to booleans
on a answer submit.

* reducer state and initialization

to avoid an unccessary render, an initializer function is passed into
the `userReducer` which will modify the initial test data if local
storage data exists, removes the need for `useEffect.

* `Question` init

`userchoice` and `isCorrect` are now initialized with a function whose
value is based on `testData`. Modify test to wrap `Question` in the
`TestDataContextProvider` since the component now uses the context.

* result array

adjust the type of the answerSheet array to `Answer`.

* bad forward navigation

Create contitional that throws error when user attempts to jump ahead in
the quiz. Error component renders and can navigate to the home page.

* current question

remove the `nextQuestion` property in the reducer and replace it with
`currentQuestion`, making the check for forward jumps easier to
implement and understand. Replaced all instances of nextQuestion.

* backwards nav check.

User is greeted with error page when attempting to go back to an already
answered question.

* Error message on invalid question

If an invalid question number value is read, then a differnet message on
the error component.