Keyboard nav (#8)
* user choice lock

set a boolean attribute on the buttons to disabled the `userChoice` from
being set again. Vitest test created.

* useKeyboardNav hook

create hook that takes in a target node list, first node, and last node.
It will generate the `keydown` even that allows for keyboard navigation
of the given target. Implemented the hook into the `root.tsx` component.

* Question and hook option

Adjust hook options to take in a optional object which takes in another
first or last element that could be part of the list. Implement the hook
into the `Question` component.

* `Question` test

Create test that use the `.keyboard` method of `UserEvent` from
react-testing library

* question focus

when navigating with the keyboard, focus will be applied to the first
option in a question after navigating to that question.