Readme (#12)
* test-choice fix

Update the `TestChoice.tsx` component to have the button wrap around all
content, allowing it to be selected in a easier way. Update `:hover` and
other psuedo states as well.

* readme

update the readme with general information about the project
- technology used
- development methods
- conclusion

* display gif and fix minor spellig mistake

* production pipeline

setup Github action to run on the `main` branch and deploy on Vercel.